requirements.txt
streamlit
opencv-python
numpy
requests
ultralytics
easyocr
fastapi
msal
python-dotenv
pydantic